[#515] Use the plural for packages with homogeneous contents
