+++ template-compose/app/src/main/java/co/nimblehq/template/compose/ui/screens/home/HomeViewModel.kt
@@ -1,10 +1,10 @@
 package co.nimblehq.template.compose.ui.screens.home
 
 import androidx.lifecycle.viewModelScope
-import co.nimblehq.template.compose.domain.usecase.UseCase
-import co.nimblehq.template.compose.model.UiModel
-import co.nimblehq.template.compose.model.toUiModel
+import co.nimblehq.template.compose.domain.usecases.UseCase
 import co.nimblehq.template.compose.ui.base.BaseViewModel
+import co.nimblehq.template.compose.ui.models.UiModel
+import co.nimblehq.template.compose.ui.models.toUiModel
 import co.nimblehq.template.compose.util.DispatchersProvider
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.*
