+++ template-compose/app/src/test/java/co/nimblehq/template/compose/test/MockUtil.kt
@@ -1,6 +1,6 @@
 package co.nimblehq.template.compose.test
 
-import co.nimblehq.template.compose.domain.model.Model
+import co.nimblehq.template.compose.domain.models.Model
 
 object MockUtil {
