+++ template-compose/app/src/test/java/co/nimblehq/template/compose/ui/screens/home/HomeScreenTest.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import co.nimblehq.template.compose.R
-import co.nimblehq.template.compose.domain.usecase.UseCase
+import co.nimblehq.template.compose.domain.usecases.UseCase
 import co.nimblehq.template.compose.test.MockUtil
 import co.nimblehq.template.compose.ui.AppDestination
 import co.nimblehq.template.compose.ui.screens.BaseScreenTest
