+++ template-compose/app/src/test/java/co/nimblehq/template/compose/ui/screens/home/HomeViewModelTest.kt
@@ -1,10 +1,10 @@
 package co.nimblehq.template.compose.ui.screens.home
 
 import app.cash.turbine.test
-import co.nimblehq.template.compose.domain.usecase.UseCase
-import co.nimblehq.template.compose.model.toUiModel
+import co.nimblehq.template.compose.domain.usecases.UseCase
 import co.nimblehq.template.compose.test.CoroutineTestRule
 import co.nimblehq.template.compose.test.MockUtil
+import co.nimblehq.template.compose.ui.models.toUiModel
 import co.nimblehq.template.compose.util.DispatchersProvider
 import io.kotest.matchers.shouldBe
 import io.mockk.every
