--- template-compose/data/src/main/java/co/nimblehq/template/compose/data/response/Response.kt
@@ -1,12 +1,0 @@
-package co.nimblehq.template.compose.data.response
-
-import co.nimblehq.template.compose.domain.model.Model
-import com.squareup.moshi.Json
-
-data class Response(
-    @Json(name = "id") val id: Int?
-)
-
-private fun Response.toModel() = Model(id = this.id)
-
-fun List<Response>.toModels() = this.map { it.toModel() }
