+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/responses/ErrorResponse.kt
@@ -1,0 +1,11 @@
+package co.nimblehq.template.compose.data.responses
+
+import co.nimblehq.template.compose.domain.models.Error
+import com.squareup.moshi.Json
+
+data class ErrorResponse(
+    @Json(name = "message")
+    val message: String
+)
+
+internal fun ErrorResponse.toModel() = Error(message = message)
