+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/responses/Response.kt
@@ -1,0 +1,12 @@
+package co.nimblehq.template.compose.data.responses
+
+import co.nimblehq.template.compose.domain.models.Model
+import com.squareup.moshi.Json
+
+data class Response(
+    @Json(name = "id") val id: Int?
+)
+
+private fun Response.toModel() = Model(id = this.id)
+
+fun List<Response>.toModels() = this.map { it.toModel() }
