--- template-compose/data/src/main/java/co/nimblehq/template/compose/data/service/providers/ApiServiceProvider.kt
@@ -1,11 +1,0 @@
-package co.nimblehq.template.compose.data.service.providers
-
-import co.nimblehq.template.compose.data.service.ApiService
-import retrofit2.Retrofit
-
-object ApiServiceProvider {
-
-    fun getApiService(retrofit: Retrofit): ApiService {
-        return retrofit.create(ApiService::class.java)
-    }
-}
