--- template-compose/data/src/main/java/co/nimblehq/template/compose/data/service/providers/ConverterFactoryProvider.kt
@@ -1,12 +1,0 @@
-package co.nimblehq.template.compose.data.service.providers
-
-import com.squareup.moshi.Moshi
-import retrofit2.Converter
-import retrofit2.converter.moshi.MoshiConverterFactory
-
-object ConverterFactoryProvider {
-
-    fun getMoshiConverterFactory(moshi: Moshi): Converter.Factory {
-        return MoshiConverterFactory.create(moshi)
-    }
-}
