--- template-compose/data/src/main/java/co/nimblehq/template/compose/data/service/providers/RetrofitProvider.kt
@@ -1,19 +1,0 @@
-package co.nimblehq.template.compose.data.service.providers
-
-import okhttp3.OkHttpClient
-import retrofit2.Converter
-import retrofit2.Retrofit
-
-object RetrofitProvider {
-
-    fun getRetrofitBuilder(
-        baseUrl: String,
-        okHttpClient: OkHttpClient,
-        converterFactory: Converter.Factory,
-    ): Retrofit.Builder {
-        return Retrofit.Builder()
-            .baseUrl(baseUrl)
-            .addConverterFactory(converterFactory)
-            .client(okHttpClient)
-    }
-}
