+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/services/providers/ConverterFactoryProvider.kt
@@ -1,0 +1,12 @@
+package co.nimblehq.template.compose.data.services.providers
+
+import com.squareup.moshi.Moshi
+import retrofit2.Converter
+import retrofit2.converter.moshi.MoshiConverterFactory
+
+object ConverterFactoryProvider {
+
+    fun getMoshiConverterFactory(moshi: Moshi): Converter.Factory {
+        return MoshiConverterFactory.create(moshi)
+    }
+}
