--- template-compose/data/src/main/java/co/nimblehq/template/compose/data/storage/SharedPreferencesExt.kt
@@ -1,10 +1,0 @@
-package co.nimblehq.template.compose.data.storage
-
-import android.content.SharedPreferences
-
-fun SharedPreferences.execute(operation: (SharedPreferences.Editor) -> Unit) {
-    with(edit()) {
-        operation(this)
-        apply()
-    }
-}
