+++ template-compose/data/src/main/java/co/nimblehq/template/compose/data/storages/SharedPreferencesExt.kt
@@ -1,0 +1,10 @@
+package co.nimblehq.template.compose.data.storages
+
+import android.content.SharedPreferences
+
+fun SharedPreferences.execute(operation: (SharedPreferences.Editor) -> Unit) {
+    with(edit()) {
+        operation(this)
+        apply()
+    }
+}
