--- template-compose/data/src/test/java/co/nimblehq/template/compose/data/repository/RepositoryTest.kt
@@ -1,48 +1,0 @@
-package co.nimblehq.template.compose.data.repository
-
-import co.nimblehq.template.compose.data.response.toModels
-import co.nimblehq.template.compose.data.service.ApiService
-import co.nimblehq.template.compose.data.test.MockUtil
-import co.nimblehq.template.compose.domain.repository.Repository
-import io.kotest.matchers.shouldBe
-import io.mockk.coEvery
-import io.mockk.mockk
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.test.runTest
-import org.junit.Before
-import org.junit.Test
-
-@ExperimentalCoroutinesApi
-class RepositoryTest {
-
-    private lateinit var mockService: ApiService
-    private lateinit var repository: Repository
-
-    @Before
-    fun setUp() {
-        mockService = mockk()
-        repository = RepositoryImpl(mockService)
-    }
-
-    @Test
-    fun `When request successful, it returns success`() = runTest {
-        val expected = MockUtil.responses
-        coEvery { mockService.getResponses() } returns expected
-
-        repository.getModels().collect {
-            it shouldBe expected.toModels()
-        }
-    }
-
-    @Test
-    fun `When request failed, it returns error`() = runTest {
-        val expected = Throwable()
-        coEvery { mockService.getResponses() } throws expected
-
-        repository.getModels().catch {
-            it shouldBe expected
-        }.collect()
-    }
-}
