--- template-compose/domain/src/main/java/co/nimblehq/template/compose/domain/model/Error.kt
@@ -1,5 +1,0 @@
-package co.nimblehq.template.compose.domain.model
-
-data class Error(
-    val message: String
-)
