--- template-compose/domain/src/main/java/co/nimblehq/template/compose/domain/model/Model.kt
@@ -1,5 +1,0 @@
-package co.nimblehq.template.compose.domain.model
-
-data class Model(
-    val id: Int?
-)
