+++ template-compose/domain/src/main/java/co/nimblehq/template/compose/domain/models/Model.kt
@@ -1,0 +1,5 @@
+package co.nimblehq.template.compose.domain.models
+
+data class Model(
+    val id: Int?
+)
