+++ template-compose/domain/src/main/java/co/nimblehq/template/compose/domain/repositories/AppPreferencesRepository.kt
@@ -1,0 +1,10 @@
+package co.nimblehq.template.compose.domain.repositories
+
+import kotlinx.coroutines.flow.Flow
+
+interface AppPreferencesRepository {
+
+    fun getAppPreference(): Flow<Boolean>
+
+    suspend fun updateAppPreference(appPreferencesValue: Boolean)
+}
