+++ template-compose/domain/src/main/java/co/nimblehq/template/compose/domain/repositories/Repository.kt
@@ -1,0 +1,9 @@
+package co.nimblehq.template.compose.domain.repositories
+
+import co.nimblehq.template.compose.domain.models.Model
+import kotlinx.coroutines.flow.Flow
+
+interface Repository {
+
+    fun getModels(): Flow<List<Model>>
+}
