--- template-compose/domain/src/main/java/co/nimblehq/template/compose/domain/repository/AppPreferencesRepository.kt
@@ -1,10 +1,0 @@
-package co.nimblehq.template.compose.domain.repository
-
-import kotlinx.coroutines.flow.Flow
-
-interface AppPreferencesRepository {
-
-    fun getAppPreference(): Flow<Boolean>
-
-    suspend fun updateAppPreference(appPreferencesValue: Boolean)
-}
