--- template-compose/domain/src/main/java/co/nimblehq/template/compose/domain/repository/Repository.kt
@@ -1,9 +1,0 @@
-package co.nimblehq.template.compose.domain.repository
-
-import co.nimblehq.template.compose.domain.model.Model
-import kotlinx.coroutines.flow.Flow
-
-interface Repository {
-
-    fun getModels(): Flow<List<Model>>
-}
