+++ template-compose/domain/src/test/java/co/nimblehq/template/compose/domain/test/MockUtil.kt
@@ -1,6 +1,6 @@
 package co.nimblehq.template.compose.domain.test
 
-import co.nimblehq.template.compose.domain.model.Model
+import co.nimblehq.template.compose.domain.models.Model
 
 object MockUtil {
